+++ vignettes/Using-assess.R
@@ -4,13 +4,14 @@
 #    .only_common_factors = TRUE,
 #    .quality_criterion   = c("all", "ave", "rho_C", "rho_C_mm", "rho_C_weighted",
 #                             "rho_C_weighted_mm", "cronbachs_alpha",
-#                            "cronbachs_alpha_weighted", "dg", "dl", "dml", "df",
-#                            "esize", "cfi", "gfi", "ifi", "nfi", "nnfi",
-#                            "reliability",
-#                            "rmsea", "rms_theta", "srmr",
-#                            "gof", "htmt", "r2", "r2_adj",
-#                            "rho_T", "rho_T_weighted", "vif",
-#                            "vifmodeB",  "fl_criterion"),
+#                             "cronbachs_alpha_weighted", "dg", "dl", "dml", "df",
+#                             "esize", "chi_square", "chi_square_df",
+#                             "cfi", "gfi", "ifi", "nfi", "nnfi",
+#                             "reliability",
+#                             "rmsea", "rms_theta", "srmr",
+#                             "gof", "htmt", "r2", "r2_adj",
+#                             "rho_T", "rho_T_weighted", "vif",
+#                             "vifmodeB",  "fl_criterion"),
 #    ...
 #  )